requirements.txt
# Librerías principales del proyecto
Django==5.2.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-filter==23.5
django-guardian==2.4.0
django-cors-headers==4.3.1
PyMySQL==1.1.1
cryptography==42.0.8
xlsxwriter==3.1.9
whitenoise==6.6.0
Pillow==10.4.0
python-decouple==3.8

# Dependencias implícitas de Django y librerías usadas
pytz==2024.2
sqlparse==0.5.1
asgiref==3.8.1
setuptools>=65.0.0
wheel>=0.41.0

# Dependencias de xhtml2pdf (REMOVIDAS: no se usan y requieren compilación en Windows)

# Dependencias opcionales pero recomendadas
six==1.16.0

# Dependencias adicionales para Docker y producción
gunicorn==21.2.0
redis==5.0.1
celery==5.3.4
django-celery-beat==2.8.0
django-celery-results==2.5.1
